Cleanup: move explorer files to one folder
