+++ charts/covidDataExplorer/CovidExplorerTable.ts
@@ -36,7 +36,7 @@ import {
     Row,
     generateEntityId
 } from "charts/owidData/OwidTable"
-import { CovidConstrainedQueryParams, CovidQueryParams } from "./CovidChartUrl"
+import { CovidConstrainedQueryParams, CovidQueryParams } from "./CovidParams"
 import {
     covidDataPath,
     covidLastUpdatedPath,
