+++ charts/covidDataExplorer/CovidParams.ts
