+++ dataExplorer/client/DataExplorers.tsx
@@ -1,23 +1,23 @@
 import React from "react"
 import { observer } from "mobx-react"
-import { CommandPalette, Command } from "./CommandPalette"
+import { CommandPalette, Command } from "charts/CommandPalette"
 import classNames from "classnames"
 import { computed, action, observable, when, reaction } from "mobx"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { Bounds } from "./Bounds"
+import { Bounds } from "charts/Bounds"
 import { faChartLine } from "@fortawesome/free-solid-svg-icons/faChartLine"
-import { ChartView } from "./ChartView"
-import { CountryPicker } from "./CountryPicker"
-import { ExplorerControlBar } from "./ExplorerControls"
-import { ChartConfig, ChartConfigProps } from "./ChartConfig"
-import { throttle, uniq } from "./Util"
+import { ChartView } from "charts/ChartView"
+import { CountryPicker } from "charts/CountryPicker"
+import { ExplorerControlBar } from "dataExplorer/client/ExplorerControls"
+import { ChartConfig, ChartConfigProps } from "charts/ChartConfig"
+import { throttle, uniq } from "charts/Util"
 import { SwitcherOptions } from "./SwitcherOptions"
-import { ExplorerControlPanel } from "./ExplorerControls"
+import { ExplorerControlPanel } from "dataExplorer/client/ExplorerControls"
 import {
     ChartQueryParams,
     ExtendedChartUrl,
     EntityUrlBuilder
-} from "./ChartUrl"
+} from "charts/ChartUrl"
 import ReactDOM from "react-dom"
 import { UrlBinder } from "charts/UrlBinder"
 import { QueryParams } from "utils/client/url"
