+++ site/client/owid.scss
@@ -37,9 +37,9 @@
 @import "css/pages/donate.scss";
 @import "css/pages/faq.scss";
 @import "css/pages/commandPalette.scss";
-@import "css/pages/dataExplorer.scss";
+@import "../../dataExplorer/client/dataExplorer.scss";
+@import "../../dataExplorer/client/explorerControls.scss";
 @import "css/pages/covidDataExplorer.scss";
-@import "css/pages/explorerControls.scss";
 @import "css/pages/countryPicker.scss";
 @import "css/pages/covid.scss";
 @import "css/pages/chart.scss";
