--- stories/co2Explorer.test.ts
@@ -1,7153 +1,0 @@
-export const co2ChartConfigs = [
-    {
-        id: 488,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "Reds",
-                isManualBuckets: true,
-                colorSchemeInvert: false,
-                colorSchemeLabels: [
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    ""
-                ],
-                colorSchemeValues: [
-                    50000000,
-                    100000000,
-                    250000000,
-                    500000000,
-                    1000000000,
-                    2500000000,
-                    5000000000,
-                    7500000000,
-                    10000000000
-                ],
-                legendDescription: "Annual CO₂ emissions",
-                colorSchemeMinValue: 0,
-                customNumericColors: [],
-                customCategoryColors: [],
-                customCategoryLabels: [],
-                customHiddenCategories: []
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142955,
-            timeTolerance: 1
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Africa",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Anguilla",
-                "Antarctic Fisheries",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Aruba",
-                "Asia (excl. China & India)",
-                "Asia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bonaire Sint Eustatius and Saba",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "British Virgin Islands",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Christmas Island",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Czechoslovakia",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "EU-27",
-                "EU-28",
-                "EU-29",
-                "EU-30",
-                "EU-31",
-                "EU-32",
-                "EU-33",
-                "EU-34",
-                "EU-35",
-                "EU-36",
-                "EU-37",
-                "EU-38",
-                "EU-39",
-                "EU-40",
-                "EU-41",
-                "EU-42",
-                "EU-43",
-                "EU-44",
-                "EU-45",
-                "EU-46",
-                "EU-47",
-                "EU-48",
-                "EU-49",
-                "EU-50",
-                "EU-51",
-                "EU-52",
-                "EU-53",
-                "EU-54",
-                "EU-55",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Europe (excl. EU-27)",
-                "Europe (excl. EU-28)",
-                "Europe",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Equatorial Africa",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "International transport",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "KP Annex B",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kosovo",
-                "Kuwait",
-                "Kuwaiti Oil Fires",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macau",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Micronesia",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Montserrat",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands Antilles",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "Non KP Annex B",
-                "Non-OECD",
-                "North America (excl. USA)",
-                "North America",
-                "North Korea",
-                "Norway",
-                "OECD",
-                "Oceania",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Puerto Rico",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Helena",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Sint Maarten (Dutch part)",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South America",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Statistical Difference",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Timor",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Turks and Caicos Islands",
-                "Tuvalu",
-                "USSR",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United Korea",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "Wallis and Futuna Islands",
-                "World",
-                "Yemen",
-                "Yugoslavia",
-                "Zambia",
-                "Zanzibar",
-                "Zimbabwe"
-            ]
-        },
-        slug: "annual-co2-emissions-per-country",
-        type: "LineChart",
-        title: "Annual CO₂ emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: {
-            min: 0,
-            scaleType: "linear",
-            labelDistance: -1,
-            canChangeScaleType: true
-        },
-        maxTime: 2017,
-        minTime: "earliest",
-        version: 18,
-        subtitle:
-            "Annual carbon dioxide (CO₂) emissions, measured in tonnes per year.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions/",
-        stackMode: null,
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142955 }],
-        entityType: "country",
-        hideLegend: false,
-        sourceDesc:
-            "Global Carbon Project; Carbon Dioxide Information Analysis Centre (CDIAC)",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 3 }
-        ],
-        internalNotes: "",
-        addCountryMode: "add-country",
-        relatedQuestions: [
-            {
-                url: "https://ourworldindata.org/annual-co2-emissions",
-                text: "Which countries emit the most CO₂ today?"
-            }
-        ],
-        hideRelativeToggle: false,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 3219,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlGnBu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.5, 1, 2, 5, 10, 25, 50],
-                customNumericColors: [null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142962
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Africa",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Anguilla",
-                "Antarctic Fisheries",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Aruba",
-                "Asia (excl. China & India)",
-                "Asia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bonaire Sint Eustatius and Saba",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "British Virgin Islands",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Christmas Island",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Czechoslovakia",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "EU-27",
-                "EU-28",
-                "EU-29",
-                "EU-30",
-                "EU-31",
-                "EU-32",
-                "EU-33",
-                "EU-34",
-                "EU-35",
-                "EU-36",
-                "EU-37",
-                "EU-38",
-                "EU-39",
-                "EU-40",
-                "EU-41",
-                "EU-42",
-                "EU-43",
-                "EU-44",
-                "EU-45",
-                "EU-46",
-                "EU-47",
-                "EU-48",
-                "EU-49",
-                "EU-50",
-                "EU-51",
-                "EU-52",
-                "EU-53",
-                "EU-54",
-                "EU-55",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Europe (excl. EU-27)",
-                "Europe (excl. EU-28)",
-                "Europe",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Equatorial Africa",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "International transport",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "KP Annex B",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kosovo",
-                "Kuwait",
-                "Kuwaiti Oil Fires",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macau",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Micronesia",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Montserrat",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands Antilles",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "Non KP Annex B",
-                "Non-OECD",
-                "North America (excl. USA)",
-                "North America",
-                "North Korea",
-                "Norway",
-                "OECD",
-                "Oceania",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Puerto Rico",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Helena",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Sint Maarten (Dutch part)",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South America",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Statistical Difference",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Timor",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Turks and Caicos Islands",
-                "Tuvalu",
-                "USSR",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United Korea",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "Wallis and Futuna Islands",
-                "World",
-                "Yemen",
-                "Yugoslavia",
-                "Zambia",
-                "Zanzibar",
-                "Zimbabwe"
-            ]
-        },
-        slug: "annual-share-of-co2-emissions",
-        type: "LineChart",
-        title: "Annual share of global CO₂ emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: 2017,
-        minTime: 1753,
-        version: 12,
-        subtitle:
-            "Each country's share of global carbon dioxide (CO₂) emissions. This is measured as each country's emissions divided by the sum of all countries' emissions in a given year plus international aviation and shipping (known as 'bunkers') and 'statistical differences' in carbon accounts.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142962 }],
-        sourceDesc: "Our World in Data based on Global Carbon Project (2018)",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 44 },
-            { index: 0, entityId: 37 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 486,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "OrRd",
-                isManualBuckets: true,
-                colorSchemeInvert: false,
-                colorSchemeLabels: [
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    "",
-                    ""
-                ],
-                colorSchemeValues: [1, 2.5, 5, 7.5, 10, 15, 20, 50],
-                legendDescription: "CO₂ emissions per capita",
-                colorSchemeMinValue: 0,
-                customNumericColors: [null],
-                customCategoryColors: [],
-                customCategoryLabels: [],
-                customHiddenCategories: []
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142960,
-            timeTolerance: 1
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Africa",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Anguilla",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Aruba",
-                "Asia (excl. China & India)",
-                "Asia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bonaire Sint Eustatius and Saba",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "British Virgin Islands",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "EU-27",
-                "EU-28",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Europe (excl. EU-27)",
-                "Europe (excl. EU-28)",
-                "Europe",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Montserrat",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North America (excl. USA)",
-                "North America",
-                "North Korea",
-                "Norway",
-                "Oceania",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Helena",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Sint Maarten (Dutch part)",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South America",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Timor",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Turks and Caicos Islands",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        note: "",
-        slug: "co-emissions-per-capita",
-        type: "LineChart",
-        title: "Per capita CO₂ emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear", labelDistance: -10 },
-        maxTime: 2017,
-        minTime: 1946,
-        version: 20,
-        subtitle:
-            "Average carbon dioxide (CO₂) emissions per capita measured in tonnes per year.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions/",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142960 }],
-        entityType: "country",
-        hideLegend: false,
-        sourceDesc:
-            "OWID based on CDIAC; Global Carbon Project; Gapminder & UN",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 104 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 355 },
-            { index: 0, entityId: 3 },
-            { index: 0, entityId: 172 }
-        ],
-        internalNotes: "",
-        addCountryMode: "add-country",
-        relatedQuestions: [
-            {
-                url: "https://ourworldindata.org/per-capita-co2",
-                text: "Where in the world do people emit the most CO₂?"
-            }
-        ],
-        hideRelativeToggle: false,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 485,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "OrRd",
-                isManualBuckets: true,
-                colorSchemeInvert: false,
-                colorSchemeLabels: ["", "", "", "", "", "", "", "", "", "", ""],
-                colorSchemeValues: [
-                    50000000,
-                    500000000,
-                    5000000000,
-                    50000000000,
-                    100000000000,
-                    250000000000,
-                    400000000000
-                ],
-                legendDescription: "Cumulative CO₂ emissions",
-                customColorsActive: false,
-                colorSchemeMinValue: 0,
-                customNumericColors: [null],
-                customCategoryColors: [],
-                customCategoryLabels: [],
-                customHiddenCategories: []
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142964,
-            timeTolerance: 1
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Africa",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Anguilla",
-                "Antarctic Fisheries",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Aruba",
-                "Asia (excl. China & India)",
-                "Asia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bonaire Sint Eustatius and Saba",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "British Virgin Islands",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Christmas Island",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Czechoslovakia",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "EU-27",
-                "EU-28",
-                "EU-29",
-                "EU-30",
-                "EU-31",
-                "EU-32",
-                "EU-33",
-                "EU-34",
-                "EU-35",
-                "EU-36",
-                "EU-37",
-                "EU-38",
-                "EU-39",
-                "EU-40",
-                "EU-41",
-                "EU-42",
-                "EU-43",
-                "EU-44",
-                "EU-45",
-                "EU-46",
-                "EU-47",
-                "EU-48",
-                "EU-49",
-                "EU-50",
-                "EU-51",
-                "EU-52",
-                "EU-53",
-                "EU-54",
-                "EU-55",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Europe (excl. EU-27)",
-                "Europe (excl. EU-28)",
-                "Europe",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Equatorial Africa",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "International transport",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "KP Annex B",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kosovo",
-                "Kuwait",
-                "Kuwaiti Oil Fires",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macau",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Micronesia",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Montserrat",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands Antilles",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "Non KP Annex B",
-                "Non-OECD",
-                "North America (excl. USA)",
-                "North America",
-                "North Korea",
-                "Norway",
-                "OECD",
-                "Oceania",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Puerto Rico",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Helena",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Sint Maarten (Dutch part)",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South America",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Statistical Difference",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Timor",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Turks and Caicos Islands",
-                "Tuvalu",
-                "USSR",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United Korea",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "Wallis and Futuna Islands",
-                "World",
-                "Yemen",
-                "Yugoslavia",
-                "Zambia",
-                "Zanzibar",
-                "Zimbabwe"
-            ]
-        },
-        note: "",
-        slug: "cumulative-co-emissions",
-        type: "LineChart",
-        title: "Cumulative CO₂ emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear", labelDistance: 0 },
-        maxTime: 2017,
-        minTime: "earliest",
-        version: 10,
-        subtitle:
-            "Cumulative carbon dioxide (CO₂) emissions represents the total sum of CO₂ emissions produced from fossil fuels and cement since 1751, and is measured in tonnes.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions/",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142964 }],
-        entityType: "country",
-        hideLegend: false,
-        sourceDesc:
-            "Global Carbon Project (GCP); Carbon Dioxide Information Analysis Centre (CDIAC)",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 3 },
-            { index: 0, entityId: 37 }
-        ],
-        internalNotes: "",
-        addCountryMode: "add-country",
-        hideRelativeToggle: false,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 3218,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlGnBu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.5, 1, 2.5, 5, 10, 25, 50],
-                customNumericColors: [null, null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142965
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Africa",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Anguilla",
-                "Antarctic Fisheries",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Aruba",
-                "Asia (excl. China & India)",
-                "Asia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bonaire Sint Eustatius and Saba",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "British Virgin Islands",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Christmas Island",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Czechoslovakia",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "EU-27",
-                "EU-28",
-                "EU-29",
-                "EU-30",
-                "EU-31",
-                "EU-32",
-                "EU-33",
-                "EU-34",
-                "EU-35",
-                "EU-36",
-                "EU-37",
-                "EU-38",
-                "EU-39",
-                "EU-40",
-                "EU-41",
-                "EU-42",
-                "EU-43",
-                "EU-44",
-                "EU-45",
-                "EU-46",
-                "EU-47",
-                "EU-48",
-                "EU-49",
-                "EU-50",
-                "EU-51",
-                "EU-52",
-                "EU-53",
-                "EU-54",
-                "EU-55",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Europe (excl. EU-27)",
-                "Europe (excl. EU-28)",
-                "Europe",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Equatorial Africa",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "International transport",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "KP Annex B",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kosovo",
-                "Kuwait",
-                "Kuwaiti Oil Fires",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macau",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Micronesia",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Montserrat",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands Antilles",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "Non KP Annex B",
-                "Non-OECD",
-                "North America (excl. USA)",
-                "North America",
-                "North Korea",
-                "Norway",
-                "OECD",
-                "Oceania",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Puerto Rico",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Helena",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Sint Maarten (Dutch part)",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South America",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Statistical Difference",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Timor",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Turks and Caicos Islands",
-                "Tuvalu",
-                "USSR",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United Korea",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "Wallis and Futuna Islands",
-                "World",
-                "Yemen",
-                "Yugoslavia",
-                "Zambia",
-                "Zanzibar",
-                "Zimbabwe"
-            ]
-        },
-        slug: "share-of-cumulative-co2",
-        type: "LineChart",
-        title: "Share of global cumulative CO₂ emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: 2017,
-        minTime: 1800,
-        version: 7,
-        subtitle:
-            "Each country or region's share of cumulative global carbon dioxide (CO₂) emissions. Cumulative emissions are calculated as the sum of annuals emissions from 1751 to a given year.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142965 }],
-        entityType: "country/region",
-        sourceDesc: "OWID based on CDIAC & Global Carbon Project (GCP)",
-        hasChartTab: true,
-        isAutoTitle: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 3 },
-            { index: 0, entityId: 81 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 4267,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlGnBu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.1, 0.2, 0.3, 0.4, 0.5, 0.6],
-                customNumericColors: [null, null, null, null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142968,
-            timeTolerance: 2
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Benin",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Bulgaria",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Costa Rica",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Libya",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Malaysia",
-                "Malta",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Panama",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Saint Lucia",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "co2-per-unit-energy",
-        type: "LineChart",
-        title: "Carbon intensity of energy production",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 10,
-        subtitle:
-            "Carbon intensity of energy production is measured as the quantity of carbon dioxide emitted per unit of energy production. This is measured in kilograms of CO₂ per kilowatt-hour.",
-        hasMapTab: true,
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [
-            {
-                display: {
-                    name: "CO₂ per unit energy (kgCO2 per kg oil equivalent)",
-                    unit: "kilograms CO₂ per kilowatt-hour",
-                    shortUnit: "kg/kWh"
-                },
-                property: "y",
-                variableId: 142968
-            }
-        ],
-        sourceDesc:
-            "Calculated by Our World in Data based on Global Carbon Project; BP; IEA via the World Bank",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 2 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 3 },
-            { index: 0, entityId: 10 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true
-    },
-    {
-        id: 530,
-        map: {
-            colorScale: {
-                baseColorScheme: "OrRd",
-                isManualBuckets: true,
-                colorSchemeInvert: false,
-                colorSchemeLabels: ["", "", "", "", "", "", "", "", "", ""],
-                colorSchemeValues: [
-                    0.1,
-                    0.2,
-                    0.3,
-                    0.4,
-                    0.5,
-                    0.6,
-                    0.8,
-                    1,
-                    1.5,
-                    2.5
-                ],
-                legendDescription: "CO2 intensity (kgCO2 per unit GDP)",
-                customColorsActive: false,
-                colorSchemeMinValue: 0,
-                customNumericColors: [],
-                customCategoryColors: [],
-                customCategoryLabels: [],
-                customHiddenCategories: []
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142966,
-            timeTolerance: 1
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Africa",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Argentina",
-                "Armenia",
-                "Asia (excl. China & India)",
-                "Asia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Benin",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Czechoslovakia",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "EU-27",
-                "EU-28",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Europe (excl. EU-27)",
-                "Europe (excl. EU-28)",
-                "Europe",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North America (excl. USA)",
-                "North America",
-                "North Korea",
-                "Norway",
-                "Oceania",
-                "Oman",
-                "Pakistan",
-                "Palestine",
-                "Panama",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "South Africa",
-                "South America",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "co2-intensity",
-        type: "LineChart",
-        title: "Carbon emission intensity of economies",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear", labelDistance: -10 },
-        maxTime: "latest",
-        minTime: 1870,
-        version: 5,
-        subtitle:
-            "Carbon dioxide (CO₂) intensity of economies measured in kilograms of CO₂ per unit of GDP (measured in international-$ in 2011 prices).",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions/",
-        stackMode: null,
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142966 }],
-        entityType: "country",
-        hideLegend: false,
-        sourceDesc:
-            "Our World in Data based on Global Carbon Project; CDIAC; and Maddison Project Database",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 355 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 6 }
-        ],
-        internalNotes: "",
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 3621,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "Reds",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    50000000,
-                    100000000,
-                    250000000,
-                    500000000,
-                    1000000000,
-                    2500000000,
-                    5000000000,
-                    10000000000
-                ],
-                customNumericColors: [null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142957
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Albania",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahrain",
-                "Bangladesh",
-                "Belarus",
-                "Belgium",
-                "Benin",
-                "Bolivia",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Chile",
-                "China",
-                "Colombia",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cyprus",
-                "Czech Republic",
-                "Denmark",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Estonia",
-                "Ethiopia",
-                "Finland",
-                "France",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Guatemala",
-                "Guinea",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "India",
-                "Indonesia",
-                "International transport",
-                "Iran",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "KP Annex B",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lithuania",
-                "Luxembourg",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Malta",
-                "Mauritius",
-                "Mexico",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Nigeria",
-                "Non KP Annex B",
-                "Non-OECD",
-                "Norway",
-                "OECD",
-                "Oman",
-                "Pakistan",
-                "Panama",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saudi Arabia",
-                "Senegal",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Statistical Difference",
-                "Sweden",
-                "Switzerland",
-                "Taiwan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Venezuela",
-                "Vietnam",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "consumption-co2-emissions",
-        type: "LineChart",
-        title: "Consumption-based CO₂ emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1990,
-        version: 9,
-        subtitle:
-            "In contrast to production-based emissions, consumption-based emissions are adjusted for trade: emissions that were caused abroad by the production of imported goods are included; emissions caused in the production of goods that were exported to another country are excluded.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142957 }],
-        sourceDesc: "Le Quéré et al. (2019). Global Carbon Project.",
-        hasChartTab: true,
-        isAutoTitle: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 10 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 81 },
-            { index: 0, entityId: 113 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: false,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 3488,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "Reds",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [1, 2.5, 5, 7.5, 10, 15, 20, 50],
-                customNumericColors: [null, null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142961
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Albania",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahrain",
-                "Bangladesh",
-                "Belarus",
-                "Belgium",
-                "Benin",
-                "Bolivia",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Chile",
-                "China",
-                "Colombia",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cyprus",
-                "Czech Republic",
-                "Denmark",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Estonia",
-                "Ethiopia",
-                "Finland",
-                "France",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Guatemala",
-                "Guinea",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lithuania",
-                "Luxembourg",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Malta",
-                "Mauritius",
-                "Mexico",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Nigeria",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Panama",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saudi Arabia",
-                "Senegal",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sweden",
-                "Switzerland",
-                "Taiwan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Venezuela",
-                "Vietnam",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "consumption-co2-per-capita",
-        type: "LineChart",
-        title: "Per capita consumption-based CO₂ emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1990,
-        version: 8,
-        subtitle:
-            "Consumption-based carbon dioxide (CO₂) emissions are national or regional emissions which have been adjusted for trade (i.e. territorial/production emissions minus emissions embedded in exports, plus emissions embedded in imports).",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142961 }],
-        sourceDesc: "OWID based on Global Carbon Project & UN Population",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 28 },
-            { index: 0, entityId: 4 },
-            { index: 0, entityId: 23 },
-            { index: 0, entityId: 10 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: false
-    },
-    {
-        map: {
-            colorScale: {
-                baseColorScheme: "YlOrRd",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8],
-                customNumericColors: [null, null, null, null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142967
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Albania",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahrain",
-                "Bangladesh",
-                "Belarus",
-                "Belgium",
-                "Benin",
-                "Bolivia",
-                "Botswana",
-                "Brazil",
-                "Bulgaria",
-                "Burkina Faso",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Chile",
-                "China",
-                "Colombia",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cyprus",
-                "Czech Republic",
-                "Denmark",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Estonia",
-                "Ethiopia",
-                "Finland",
-                "France",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Guatemala",
-                "Guinea",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lithuania",
-                "Luxembourg",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Malta",
-                "Mauritius",
-                "Mexico",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Nigeria",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Panama",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saudi Arabia",
-                "Senegal",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sweden",
-                "Switzerland",
-                "Taiwan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Venezuela",
-                "Vietnam",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        note:
-            "For consumption-based emissions,  emissions that were caused abroad by the production of imported goods are included; emissions caused in the production of goods that were exported to another country are excluded.",
-        slug: "consumption-based-carbon-intensity",
-        type: "LineChart",
-        title: "Consumption-based carbon intensity",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 2,
-        subtitle:
-            "Carbon intensity measured in kilograms of CO₂ per unit of GDP (measured in international-$ in 2011 prices). This is measured based on consumption-based CO₂ emissions.",
-        hasMapTab: true,
-        originUrl: "ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142967 }],
-        isAutoSlug: true,
-        sourceDesc:
-            "Our World in Data based on Global Carbon Project and World Bank",
-        hasChartTab: true,
-        isAutoTitle: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 10 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 6 },
-            { index: 0, entityId: 171 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true,
-        id: 4331
-    },
-    {
-        id: 696,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "RdBu",
-                isManualBuckets: true,
-                colorSchemeInvert: true,
-                colorSchemeLabels: ["", "", "", "", "", "", "", "", "", ""],
-                colorSchemeValues: [
-                    -60,
-                    -40,
-                    -20,
-                    -10,
-                    0,
-                    10,
-                    20,
-                    50,
-                    100,
-                    450
-                ],
-                legendDescription:
-                    "CO₂ traded as share of domestic (production) emissions",
-                customColorsActive: false,
-                colorSchemeMinValue: -80,
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142959,
-            timeTolerance: 1
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Albania",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahrain",
-                "Bangladesh",
-                "Belarus",
-                "Belgium",
-                "Benin",
-                "Bolivia",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Chile",
-                "China",
-                "Colombia",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cyprus",
-                "Czech Republic",
-                "Denmark",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Estonia",
-                "Ethiopia",
-                "Finland",
-                "France",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Guatemala",
-                "Guinea",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "India",
-                "Indonesia",
-                "International transport",
-                "Iran",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "KP Annex B",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lithuania",
-                "Luxembourg",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Malta",
-                "Mauritius",
-                "Mexico",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Nigeria",
-                "Non KP Annex B",
-                "Non-OECD",
-                "Norway",
-                "OECD",
-                "Oman",
-                "Pakistan",
-                "Panama",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saudi Arabia",
-                "Senegal",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Statistical Difference",
-                "Sweden",
-                "Switzerland",
-                "Taiwan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Venezuela",
-                "Vietnam",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        note: "",
-        slug: "share-co2-embedded-in-trade",
-        type: "LineChart",
-        title: "CO₂ emissions embedded in trade",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, suffix: "%", scaleType: "linear", labelDistance: -5 },
-        maxTime: 2016,
-        minTime: 1990,
-        version: 8,
-        subtitle:
-            'Share of carbon dioxide (CO₂) emissions embedded in trade, measured as emissions exported or imported as the percentage of domestic production emissions. Positive values (red) represent net importers of CO₂ (i.e. "20%" would mean a country imported emissions equivalent to 20% of its domestic emissions). Negative values (blue) represent net exporters of CO₂.',
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions/",
-        stackMode: null,
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142959 }],
-        entityType: "country",
-        hideLegend: false,
-        sourceDesc:
-            "Peters et al. (2012 updated); Global Carbon Project (2018)",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 4 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 10 }
-        ],
-        internalNotes: "",
-        addCountryMode: "add-country",
-        hideRelativeToggle: true
-    },
-    {
-        id: 4250,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "Oranges",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    1000000,
-                    5000000,
-                    10000000,
-                    50000000,
-                    100000000,
-                    500000000,
-                    1000000000,
-                    5000000000
-                ],
-                colorSchemeMinValue: 0,
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 104492
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "annual-co2-coal",
-        type: "LineChart",
-        title: "Annual CO₂ emissions from coal",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 4,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [
-            {
-                display: { numDecimalPlaces: 0 },
-                property: "y",
-                variableId: 104492
-            }
-        ],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 81 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 4251,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "PuBu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    10000000,
-                    50000000,
-                    100000000,
-                    50000000,
-                    1000000000,
-                    5000000000
-                ],
-                colorSchemeMinValue: 0,
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 104493
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Anguilla",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Aruba",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bonaire Sint Eustatius and Saba",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "British Virgin Islands",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Montserrat",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Helena",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Timor",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Turks and Caicos Islands",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "annual-co2-oil",
-        type: "LineChart",
-        title: "Annual CO₂ emissions from oil",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1900,
-        version: 6,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 104493 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 90 },
-            { index: 0, entityId: 103 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 37 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlOrRd",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    10000000,
-                    50000000,
-                    100000000,
-                    500000000,
-                    1000000000,
-                    5000000000
-                ],
-                colorSchemeMinValue: 0,
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 104491
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zimbabwe"
-            ]
-        },
-        slug: "annual-co2-gas",
-        type: "LineChart",
-        title: "Annual CO₂ emissions from gas",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1900,
-        version: 3,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissionshttps://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 104491 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 44 },
-            { index: 0, entityId: 12 },
-            { index: 0, entityId: 102 },
-            { index: 0, entityId: 14 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true,
-        id: 4253
-    },
-    {
-        id: 4255,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "PuBuGn",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    1000000,
-                    5000000,
-                    10000000,
-                    50000000,
-                    100000000,
-                    500000000,
-                    1000000000
-                ],
-                customNumericColors: [null, null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 104494
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Liberia",
-                "Libya",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "annual-co2-cement",
-        type: "LineChart",
-        title: "Annual CO₂ emissions from cement",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1900,
-        version: 4,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 104494 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 72 },
-            { index: 0, entityId: 14 },
-            { index: 0, entityId: 81 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        map: {
-            colorScale: {
-                baseColorScheme: "YlGnBu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    5000000,
-                    10000000,
-                    15000000,
-                    20000000,
-                    25000000
-                ],
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 104495
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen"
-            ]
-        },
-        slug: "annual-co2-flaring",
-        type: "LineChart",
-        title: "Annual CO₂ emissions from flaring",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 3,
-        hasMapTab: true,
-        originUrl: "ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 104495 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 44 },
-            { index: 0, entityId: 171 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true,
-        id: 4332
-    },
-    {
-        id: 4249,
-        map: {
-            colorScale: {
-                baseColorScheme: "YlOrBr",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [1, 2, 3, 4, 5, 6],
-                customNumericColors: [null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142850
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "per-capita-co2-coal",
-        type: "LineChart",
-        title: "Per capita CO2 emissions from coal",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1860,
-        version: 5,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142850 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 2 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 23 },
-            { index: 0, entityId: 81 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        map: {
-            colorScale: {
-                baseColorScheme: "PuBu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [1, 2, 3, 4, 5, 6, 7],
-                colorSchemeMinValue: 0,
-                customNumericColors: [null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142851
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Anguilla",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Aruba",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bonaire Sint Eustatius and Saba",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "British Virgin Islands",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Montserrat",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Helena",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "South Sudan",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Timor",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Turks and Caicos Islands",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "per-capita-co2-oil",
-        type: "LineChart",
-        title: "Per capita CO2 emissions from oil",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1954,
-        version: 3,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142851 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 90 },
-            { index: 0, entityId: 208 },
-            { index: 0, entityId: 72 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true,
-        id: 4252
-    },
-    {
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlOrRd",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [2, 4, 6, 8, 10],
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142852
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Curacao",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zimbabwe"
-            ]
-        },
-        slug: "per-capita-co2-gas",
-        type: "LineChart",
-        title: "Per capita CO2 emissions from gas",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1900,
-        version: 3,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142852 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 12 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 44 },
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 172 },
-            { index: 0, entityId: 171 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true,
-        id: 4254
-    },
-    {
-        map: {
-            colorScale: {
-                baseColorScheme: "PuBuGn",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.1, 0.2, 0.3, 0.4, 0.5, 0.6],
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142853
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Liberia",
-                "Libya",
-                "Lithuania",
-                "Luxembourg",
-                "Macao",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "per-capita-co2-cement",
-        type: "LineChart",
-        title: "Per capita CO2 emissions from cement",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: 1930,
-        version: 3,
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142853 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 23 },
-            { index: 0, entityId: 81 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true,
-        id: 4256
-    },
-    {
-        map: {
-            colorScale: {
-                baseColorScheme: "GnBu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.05, 0.1, 0.15, 0.2, 0.25, 0.3],
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142854
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bermuda",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Cayman Islands",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Estonia",
-                "Ethiopia",
-                "Faeroe Islands",
-                "Falkland Islands",
-                "Fiji",
-                "Finland",
-                "France",
-                "French Guiana",
-                "French Polynesia",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Gibraltar",
-                "Greece",
-                "Greenland",
-                "Grenada",
-                "Guadeloupe",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hong Kong",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malaysia",
-                "Mali",
-                "Malta",
-                "Martinique",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Nepal",
-                "Netherlands",
-                "New Caledonia",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Palestine",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Reunion",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Lucia",
-                "Saint Pierre and Miquelon",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Taiwan",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen"
-            ]
-        },
-        slug: "per-capita-co2-flaring",
-        type: "LineChart",
-        title: "Per capita CO2 emissions from flaring",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 3,
-        hasMapTab: true,
-        originUrl: "ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142854 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 44 },
-            { index: 0, entityId: 23 },
-            { index: 0, entityId: 171 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true,
-        id: 4333
-    },
-    {
-        id: 4147,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlOrBr",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    10000000,
-                    50000000,
-                    100000000,
-                    500000000,
-                    1000000000,
-                    5000000000,
-                    10000000000
-                ],
-                customNumericColors: [null, null, null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: 2016,
-            variableId: 142731
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "European Union (27)",
-                "Fiji",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Grenada",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Micronesia",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "total-ghg-emissions",
-        type: "LineChart",
-        title: "Total greenhouse gas emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear", canChangeScaleType: true },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 4,
-        subtitle:
-            "Greenhouse gas emissions are measured in tonnes of carbon dioxide-equivalents (CO₂e), and include emissions from land-use change and forestry.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142731 }],
-        sourceDesc: "CAIT Climate Data Explorer via. Climate Watch",
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 37 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: false,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 4239,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlOrRd",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [2.5, 5, 7.5, 10, 15, 20, 25, 50],
-                colorSchemeMinValue: 0,
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142834
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Fiji",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Grenada",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "per-capita-ghg-emissions",
-        type: "LineChart",
-        title: "Per capita greenhouse gas emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 4,
-        subtitle:
-            "Per capita greenhouse gas emissions are measured in tonnes of carbon-dioxide equivalents (CO<sub>2</sub>e) per person per year. This metric converts all greenhouse gases to  CO<sub>2</sub>e based on their global warming potential value over a 100-year timescale.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142834 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 355 },
-            { index: 0, entityId: 6 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 4222,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "OrRd",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    10000000,
-                    50000000,
-                    100000000,
-                    50000000,
-                    100000000,
-                    500000000,
-                    1000000000
-                ],
-                customNumericColors: [null],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142803
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "European Union (27)",
-                "Fiji",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Grenada",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Micronesia",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        note: "Emissions from land use change and forestry are included.",
-        slug: "methane-emissions",
-        type: "LineChart",
-        title: "Methane emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 3,
-        subtitle:
-            "Methane (CH₄) emissions are measured in tonnes of carbon dioxide equivalents (CO₂e) based on a 100-year global warming potential value. ",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142803 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 5 },
-            { index: 0, entityId: 3 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 137 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 4243,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "YlOrRd",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.5, 1, 2, 4, 6, 8],
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142841
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Fiji",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Grenada",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "per-capita-methane-emissions",
-        type: "LineChart",
-        title: "Per capita methane emissions",
-        useV2: false,
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 5,
-        subtitle:
-            "Per capita methane emissions are measured in tonnes of carbon-dioxide equivalents (CO<sub>2</sub>e) per person per year. This metric converts all greenhouse gases to  CO<sub>2</sub>e based on their global warming potential value over a 100-year timescale.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142841 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 12 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 136 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 4224,
-        map: {
-            colorScale: {
-                equalSizeBins: true,
-                baseColorScheme: "BuPu",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [
-                    10000000,
-                    50000000,
-                    100000000,
-                    500000000,
-                    1000000000
-                ],
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142812
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "European Union (27)",
-                "Fiji",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Grenada",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Micronesia",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        note: "Emissions from land use change and forestry are included.",
-        slug: "nitrous-oxide-emissions",
-        type: "LineChart",
-        title: "Nitrous oxide emissions",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 4,
-        subtitle:
-            "Nitrous oxide (N₂O) emissions are measured in tonnes of carbon dioxide equivalents (CO₂e) based on a 100-year global warming potential value.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [{ display: {}, property: "y", variableId: 142812 }],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 8 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 171 },
-            { index: 0, entityId: 13 },
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 37 }
-        ],
-        addCountryMode: "add-country",
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    },
-    {
-        id: 4244,
-        map: {
-            colorScale: {
-                baseColorScheme: "Purples",
-                isManualBuckets: true,
-                colorSchemeLabels: [],
-                colorSchemeValues: [0.5, 1, 1.5, 2, 2.5],
-                customNumericColors: [],
-                customCategoryColors: {},
-                customCategoryLabels: {},
-                customHiddenCategories: {}
-            },
-            projection: "World",
-            targetYear: "latest",
-            variableId: 142848
-        },
-        tab: "map",
-        data: {
-            availableEntities: [
-                "Afghanistan",
-                "Albania",
-                "Algeria",
-                "Andorra",
-                "Angola",
-                "Antigua and Barbuda",
-                "Argentina",
-                "Armenia",
-                "Australia",
-                "Austria",
-                "Azerbaijan",
-                "Bahamas",
-                "Bahrain",
-                "Bangladesh",
-                "Barbados",
-                "Belarus",
-                "Belgium",
-                "Belize",
-                "Benin",
-                "Bhutan",
-                "Bolivia",
-                "Bosnia and Herzegovina",
-                "Botswana",
-                "Brazil",
-                "Brunei",
-                "Bulgaria",
-                "Burkina Faso",
-                "Burundi",
-                "Cambodia",
-                "Cameroon",
-                "Canada",
-                "Cape Verde",
-                "Central African Republic",
-                "Chad",
-                "Chile",
-                "China",
-                "Colombia",
-                "Comoros",
-                "Congo",
-                "Cook Islands",
-                "Costa Rica",
-                "Cote d'Ivoire",
-                "Croatia",
-                "Cuba",
-                "Cyprus",
-                "Czech Republic",
-                "Democratic Republic of Congo",
-                "Denmark",
-                "Djibouti",
-                "Dominica",
-                "Dominican Republic",
-                "Ecuador",
-                "Egypt",
-                "El Salvador",
-                "Equatorial Guinea",
-                "Eritrea",
-                "Estonia",
-                "Ethiopia",
-                "Fiji",
-                "Finland",
-                "France",
-                "Gabon",
-                "Gambia",
-                "Georgia",
-                "Germany",
-                "Ghana",
-                "Greece",
-                "Grenada",
-                "Guatemala",
-                "Guinea-Bissau",
-                "Guinea",
-                "Guyana",
-                "Haiti",
-                "Honduras",
-                "Hungary",
-                "Iceland",
-                "India",
-                "Indonesia",
-                "Iran",
-                "Iraq",
-                "Ireland",
-                "Israel",
-                "Italy",
-                "Jamaica",
-                "Japan",
-                "Jordan",
-                "Kazakhstan",
-                "Kenya",
-                "Kiribati",
-                "Kuwait",
-                "Kyrgyzstan",
-                "Laos",
-                "Latvia",
-                "Lebanon",
-                "Lesotho",
-                "Liberia",
-                "Libya",
-                "Liechtenstein",
-                "Lithuania",
-                "Luxembourg",
-                "Macedonia",
-                "Madagascar",
-                "Malawi",
-                "Malaysia",
-                "Maldives",
-                "Mali",
-                "Malta",
-                "Marshall Islands",
-                "Mauritania",
-                "Mauritius",
-                "Mexico",
-                "Moldova",
-                "Mongolia",
-                "Montenegro",
-                "Morocco",
-                "Mozambique",
-                "Myanmar",
-                "Namibia",
-                "Nauru",
-                "Nepal",
-                "Netherlands",
-                "New Zealand",
-                "Nicaragua",
-                "Niger",
-                "Nigeria",
-                "Niue",
-                "North Korea",
-                "Norway",
-                "Oman",
-                "Pakistan",
-                "Palau",
-                "Panama",
-                "Papua New Guinea",
-                "Paraguay",
-                "Peru",
-                "Philippines",
-                "Poland",
-                "Portugal",
-                "Qatar",
-                "Romania",
-                "Russia",
-                "Rwanda",
-                "Saint Kitts and Nevis",
-                "Saint Lucia",
-                "Saint Vincent and the Grenadines",
-                "Samoa",
-                "Sao Tome and Principe",
-                "Saudi Arabia",
-                "Senegal",
-                "Serbia",
-                "Seychelles",
-                "Sierra Leone",
-                "Singapore",
-                "Slovakia",
-                "Slovenia",
-                "Solomon Islands",
-                "Somalia",
-                "South Africa",
-                "South Korea",
-                "Spain",
-                "Sri Lanka",
-                "Sudan",
-                "Suriname",
-                "Swaziland",
-                "Sweden",
-                "Switzerland",
-                "Syria",
-                "Tajikistan",
-                "Tanzania",
-                "Thailand",
-                "Togo",
-                "Tonga",
-                "Trinidad and Tobago",
-                "Tunisia",
-                "Turkey",
-                "Turkmenistan",
-                "Tuvalu",
-                "Uganda",
-                "Ukraine",
-                "United Arab Emirates",
-                "United Kingdom",
-                "United States",
-                "Uruguay",
-                "Uzbekistan",
-                "Vanuatu",
-                "Venezuela",
-                "Vietnam",
-                "World",
-                "Yemen",
-                "Zambia",
-                "Zimbabwe"
-            ]
-        },
-        slug: "per-capita-nitrous-oxide",
-        type: "LineChart",
-        title: "Per capita nitrous oxide emissions",
-        xAxis: { scaleType: "linear" },
-        yAxis: { min: 0, scaleType: "linear" },
-        maxTime: "latest",
-        minTime: "earliest",
-        version: 3,
-        subtitle:
-            "Per capita nitrous oxide emissions are measured in tonnes of carbon-dioxide equivalents (CO<sub>2</sub>e) per person per year. This metric converts all greenhouse gases to  CO<sub>2</sub>e based on their global warming potential value over a 100-year timescale.",
-        hasMapTab: true,
-        originUrl:
-            "https://ourworldindata.org/co2-and-other-greenhouse-gas-emissions",
-        stackMode: "absolute",
-        colorScale: {
-            colorSchemeLabels: [],
-            colorSchemeValues: [],
-            customNumericColors: [],
-            customCategoryColors: {},
-            customCategoryLabels: {},
-            customHiddenCategories: {}
-        },
-        dimensions: [
-            {
-                display: {
-                    name: "Total including land use change",
-                    unit: "tonnes per capita",
-                    shortUnit: "t",
-                    includeInTable: true
-                },
-                property: "y",
-                variableId: 142848
-            }
-        ],
-        hasChartTab: true,
-        isPublished: true,
-        isExplorable: false,
-        selectedData: [
-            { index: 0, entityId: 1 },
-            { index: 0, entityId: 2 },
-            { index: 0, entityId: 106 },
-            { index: 0, entityId: 137 },
-            { index: 0, entityId: 37 },
-            { index: 0, entityId: 171 }
-        ],
-        hideRelativeToggle: true,
-        hideTitleAnnotation: true
-    }
-]
