+++ stories/covidDataExplorer.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { CovidDataExplorer } from "charts/covidDataExplorer/CovidDataExplorer"
 import { covidSampleRows } from "test/fixtures/CovidSampleRows"
-import { CovidQueryParams } from "charts/covidDataExplorer/CovidChartUrl"
+import { CovidQueryParams } from "charts/covidDataExplorer/CovidParams"
 
 export default {
     title: "CovidDataExplorer",
